DSBD-20: rename run -> scenario as it fits more.
