+++ README.md
@@ -3,6 +3,8 @@ The interface for executing binaries on morello board and self that returns outp
 
 # Local
 Update your local.json to contain a variable for your QEMU morello-purecap system.
+
+> Generate swagger and route files for Open API
 > changes required in order to execute shell over ssh
 > details on how to find an ip address
