+++ src/controllers/scenario/index.ts
@@ -8,18 +8,19 @@ import {
 import Logger from '../../utils/Logger'
 import { Scenario } from "../../../types"
 
-let log = Logger.child({ controller: '/run' })
+let log = Logger.child({ controller: '/scenario' });
 
 @Route('run')
-export class Run extends Controller {
+export class scenario extends Controller {
 
   @Get('{id}')
   public get(@Path() id: string): Promise<Scenario> {
-    log.info('get example controller', id)
+    log.info(`executing ${id} scenario`);
+
     return Promise.resolve({
       status: 'error',
       output: 'some error',
-    })
+    });
   }
 }
