+++ src/index.ts
@@ -21,12 +21,6 @@ app.use(json());
 RegisterRoutes(app);
 app.use(['/openapi', '/docs', '/swagger'], swaggerUI.serve, swaggerUI.setup(swaggerJson));
 
-app.get('/scenario', (req: Request, res: Response) => {
-  log = log.child({ child: 'child-example-2'})
-  log.info(req.headers);
-  res.send('TypeScript nodejs template');
-});
-
 app.listen(port, () => {
   log.info('Server is running');
 });
