+++ types/index.d.ts
@@ -1,4 +1,4 @@
-import type { Scenario } from './models/run'
+import type { Scenario } from './models/Scenario'
 
 export type {
   Scenario,
